Extract helpers into a separate file
